requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
deprecation==2.1.0
fastapi==0.95.1
Flask==2.2.3
frozenlist==1.3.3
functions-framework==3.3.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.6
packaging==23.1
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.28.2
sniffio==1.3.0
starlette==0.26.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
watchdog==3.0.0
Werkzeug==2.2.3
yarl==1.9.2
